Making server folder, moving database to server folder. Small changes to redirect
