--- database/knex.js
@@ -1,2 +1,0 @@
-const config = require('../knexfile')
-module.exports = require('knex')(config)
