+++ server/database/knex.js
@@ -1,0 +1,2 @@
+const config = require('../knexfile');
+module.exports = require('knex')(config);
