+++ server/database/models/User.js
@@ -1,6 +1,6 @@
 const bookshelf = require('../bookshelf');
 
-class User extends bookshelf.model {
+class User extends bookshelf.Model {
     get tableName() {
         return 'users'
     }
