+++ server/database/seeds/1_categories.js
